requirements.txt
openai==1.3.0
mediafire==0.6.1
selenium==4.15.0
webdriver-manager==3.8.0
requests==2.31.0
python-wordpress-xmlrpc==2.3
Pillow==10.0.0
beautifulsoup4==4.12.0
python-dotenv==1.0.0
chromedriver-autoinstaller==0.6.2